Specify release-sign-artifacts profile in release.sh
The artifacts are only signed when the `release-sign-artifacts`
profile is activated, which is normally done by the maven-release-plugin
for the release:perform task. That's not happening for some reason, so
as a workaround, we're now explicitly setting the profile when using the
release script. We're soon switching to Gradle, so no further effort will
be put in to fix this in Maven configs.

This patch also allows passing mvn arguments to release.sh, which is
useful for setting -DdryRun=true.